Remove keybinding to open hard-coded file
I realized, fool that I am, that I could have my "nvim/init.vim" just be
a symbolic link to my init.vim in my directory of dotfiles. After all,
that's what I do for my .zshrc. So, I can safely remove the keybinding
that would open that, and instead open `$MYVIMRC`, which of course now
points to a file which is a symbolic link, which vim will happily
respect.