Add marco/polo script pair
Added a functionality that was suggested in the "missing semester of
your CS education" exercises (you can find the website at:
https://missing.csail.mit.edu/). Namely, calling "marco" in a directory
stores that path. Later, you can call ". polo" (I've aliased it in this
same commit to just "polo") to cd to that saved directory.

Note: I'm pretty sure that this functionality exists, and it in many
ways overlaps with things like pushd/popd and cd -, but 1) there are
many like it but this one is mine, and 2) I had a good time hacking away
at some of the more nuanced challenges here.